Implemented neater argument parsing + refactored into separate files + added docstrings
